Fix diagnostics CI contracts
